Add test-run-1 (blinded 2-condition savings framing study) and fix two bugs
- Fix persona_prompt.py syntax error (escaped quote inside f-string) that
  broke build_batches.py
- Fix balance_check.py Bartlett's test failing on integer-dtype columns
  (cast to float before scipy.stats.bartlett)
- Add round support to sample_population.py continuous variables, for
  ordinal/count-scale variables (income bracket, financial literacy,
  subjective numeracy)
- Add reasoning_prompt override to persona_prompt.py so the free-text
  reasoning field can match a specific instrument's wording
- First full pipeline run: N=20 synthetic subjects, blinded condition_1
  vs condition_2 (savings option framed as $5/day vs $150/month), full
  Stage 3-5 design/config/data/analysis artifacts under projects/test-run-1

+++ scripts/analysis/balance_check.py
@@ -33,7 +33,9 @@ def _categorical_balance(df, var, condition_col):
 
 
 def _continuous_balance(df, var, condition_col):
-    groups = [g[var].dropna().values for _, g in df.groupby(condition_col)]
+    # cast to float: some scipy versions (e.g. bartlett in scipy 1.18's array-api
+    # backend) fail on integer-dtype input with an internal NaN-assignment error
+    groups = [g[var].dropna().astype(float).values for _, g in df.groupby(condition_col)]
     f_stat, f_p = stats.f_oneway(*groups)
     bart_stat, bart_p = stats.bartlett(*groups)
     desc = df.groupby(condition_col)[var].agg(["mean", "std", "count"]).round(3)
